Response
Comfort: 1 / 5
Clarity: 1 / 5

I don't think my lack of comfort and clarity has to do with the
consultants not teaching it well, I think it's just the lack of time
spent on the subject. We went through those quite fast and I wish I had
more time in the evenings to review.

I think there is also some confusion caused by the fat arrow If
statements because I barely got to use and understand the If statement
syntax but we were already on to the =>. It's cool and quicker to write,
but would be nice if the callbacks were taught with the standard If
statement syntax and maybe later when we feel comfortable, teach it with
 =>'s.'