complete js array iteration methods diagnostic
Comfort: 3
Clarity: 2
I feel the way in which this material was presented confused many of us.
While I understand the reasoning for wanting to show us how these array
methods work, I think it is more important at this point for us to know
how to use them. Introducing them to us along with fat arrow formulas
led to a lot of confusion over syntax, which distracted from the real
material being presented.