Commit for js-array-iteration methods
I put in some placeholder code on a couple of the questions
that I would then work with to try and solve them.
I am not fast enough with this particular material yet to do much
more than that in a half hour. I still need to practice starting
from scratch on this material. We are often shown where to start.
I will do extra practice today.